client/postcurve_api/requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-cors-headers==2.4.0
django-log-request-id==1.3.2
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-yasg==1.13.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.0
PyJWT==1.7.1
pytz==2018.9
redis==3.2.0
requests==2.21.0
ruamel.yaml==0.15.88
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
